Add password matching checking functionality
For some reason, I had to add an event listener to the Document,
listening for a DOMContentLoaded event. Then I nested the event listener
for the submit button within that event listener. The script tag is at the
end of the body in the HTML, so I'm not sure why it wasn't working without
the event listener nested. However, when I went with this structure,
it fixed a bug where that validation was not working (would send you through
even if the password fields did not match).